refactor(web): extract static frontend from web.py into ev/interfaces/web package
ev/interfaces/web.py had grown to 6364 lines mixing the static frontend
(inline HTML/CSS/JS page, favicon, service worker, PNG icon renderer) with
the FastAPI app factory and ~177 routes. Split it into a package so the
frontend content (now ev/interfaces/web/frontend.py, ~3700 lines, pure/no
project imports) is separate from the backend routing logic (now
ev/interfaces/web/app.py, unchanged except import depth). No behavior
changed — this is a literal extract-and-recompose, part 6a of the
ongoing interfaces refactor (routes themselves are 6b, a separate PR).

ev/interfaces/web/__init__.py re-exports create_app and run so existing
imports (e.g. `from ev.interfaces.web import create_app`) keep working.
